+++ src/error.rs
@@ -1,18 +1,17 @@
-use std::fmt::{self, Display, Formatter};
 use std::io;
 
 use anyhow::{bail, Context, Result};
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+#[error("could not find fzf, is it installed?")]
+pub struct FzfNotFound;
 
 /// Custom error type for early exit.
-#[derive(Debug)]
+#[derive(Debug, Error)]
+#[error("")]
 pub struct SilentExit {
     pub code: i32,
-}
-
-impl Display for SilentExit {
-    fn fmt(&self, _: &mut Formatter) -> fmt::Result {
-        Ok(())
-    }
 }
 
 pub trait BrokenPipeHandler {